requirements.txt
asgiref==3.7.2
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-import-export==3.3.7
et-xmlfile==1.1.0
iniconfig==2.0.0
MarkupPy==1.14
mysql-connector==2.2.9
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.1
pillow==10.2.0
pluggy==1.5.0
PyMySQL==1.1.0
pytest==8.3.2
pytest-django==4.8.0
python-slugify==8.0.4
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2024.1
xlrd==2.0.1
xlwt==1.3.0
reportlab>=3.6.0
